+++ internal/agent/agent.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/tinkerbell/tink/internal/agent/event"
-	"github.com/tinkerbell/tink/internal/agent/workflow"
+	"github.com/tinkerbell/tink/internal/workflow"
 )
 
 // Agent is the core data structure for handling workflow execution on target nodes. It leverages
