+++ internal/agent/agent_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/tinkerbell/tink/internal/agent/failure"
 	"github.com/tinkerbell/tink/internal/agent/runtime"
 	"github.com/tinkerbell/tink/internal/agent/transport"
-	"github.com/tinkerbell/tink/internal/agent/workflow"
+	"github.com/tinkerbell/tink/internal/workflow"
 	"go.uber.org/zap"
 )
