+++ internal/agent/transport/fake.go
@@ -4,8 +4,9 @@ import (
 	"context"
 
 	"github.com/go-logr/logr"
+
 	"github.com/tinkerbell/tink/internal/agent/event"
-	"github.com/tinkerbell/tink/internal/agent/workflow"
+	"github.com/tinkerbell/tink/internal/workflow"
 )
 
 func Noop() Fake {
