+++ internal/controller/manager.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/tinkerbell/tink/api/v1alpha1"
-	"github.com/tinkerbell/tink/internal/workflow"
+	"github.com/tinkerbell/tink/internal/obsolete/workflow"
 	"k8s.io/apimachinery/pkg/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
