+++ internal/obsolete/workflow/template_funcs.go
